Reworked timestamps, signal handlers, and profiling.
Instead of maintaining the global 'g_now' indirectly from signal handlers just
call gettimeofday(). This results in a small performance loss, but enables us
to get rid of the 1000/sec signal handler that is constantly eating
approcimately 2-3% CPU on new machines.

The overhead of gettimeofday() since linux kernel 2.6 is lowered beacuse it is a
vsyscall.

For profiling the dedicated SIGPROF signal is now used.

SIGALARM is no longer used.

SIGVTALARM is currently disabled because it is only needed for setting the
g_loop.m_needsToQuickPoll variable, which in this fork is non-functional and
unneeded. It may be revided in the future if it turns out quickpolls are needed.