Bugfix operator new[] and delete[]
There was confusion about whether it is the compiler's or the library's task to
keep track of number of objects in a C++ array. It is the compiler's. And the
library must only keep track of memory. The code was confused and was off by 4
bytes when using C++ array new/delete.