Fix use of uninitialized data
If the query was missing when getting a cached page then code would highlight
using random words from an uninitialized buffer.  When searching for a title to
put at the top of the table-embedding of cached pages it would search for <title>
in the allocaced buffer size and not the used buffer size.