Remove Dates.*
The functionlity in Dates.* was used for extracting dates and times from
unstructured webpages for discovering theater/concert times and similar.
However, we don't need this functiolity and as far as Ii can tell it was used
for the now-defunct sites flurbit and eventguru. The only known regression is
that the sentence detection may now detect ".... at 7 .a.m. Thursday ..." as two
sentences instead of one. But that is language- and country-specific anyway.