Removed hand-coded memcpy(), memset(), etc.
The rationale for the hand-coded versions was that the functions were not
guaranteed to be async-signal-safe and since itimer() and signals are being used
then it was necessary to provide own versions of it.
My guess is that the programmer knew the standards, but did't consider that the
environment in which the program would be used in would not have a problem with
this. Besides, memcpy, memset, etc. will be marked async-signal-safe in the next
IEEE Std 1003.1. See http://austingroupbugs.net/view.php?id=692#c1609 for
details. In particular this fragment "there are no known implementations where
these functions cannot be safely used in aysnc-signal-safe code."