Removed Mem::setStackPointer and Mem::checkStackSize()
checkStackSize() was not called. The only lingering use (commneted-out) was a
check for a hardcoded stack use of 6MB. The correct way would have been to use
the API to find the stack size, and also to cater for threads that have their
own stacks. But it was uncalled so the two methods and associated members have
been removed.