Make include order more consistent
With glibc-2.18 it is possible to end up with int32_t typedef'ed as "int" in one
module and "unsigned int" in another, making the use of it nominally
inconsistent, which Flexelint complains about. But using flexelint's -fdc flag
reveals that the C++ template numeric_limits<> relies on "char", "signed char"
and "unsigned char" being three distinct types. glibc-2.18 contains at least
three different typedefs for the integer types, so it is a mess. Best option is
to make include order consistent in our source code.