Removed global variable g_inMemFunction
The updates to it were incomplete (eg more functions in openssl calls malloc(). Since we have disabled quickpolls the only two cases that remain are signal handlers (sigsegv/sigbus/...) and the profiler signal handler. For regular signal handlers we just take a chance. For profiler it only calls backtrace() which does use heap.