Make Msg5 interpret null callback as "don't create jobs or other asynchronous work"
If allback aprameter to Msg5::getList is NULL then Msg5 takes that as an order that it must not start jobs or other asynchronous jobs and return false. This is followed through via Msg3, RdbScan and BigFile.

One side-effect is that if corrupted data is detected (in bigfile or msg3) then data from a twin is not fetched. But the ode expecting synchronous calls couldn't deal with it anyway.