Use merge-space while merging
No more BigFile .part* deletion during a merge to preserve disk space. Instead MergeSpaceCoordinator is used for coordinating access to a large and possibly cheap storage with room for a whole resulting mergefile.
When a mrge file has been finished the reads are allowed from that and reads from the source files disallowed, which are then deleted. Then the file is renamed/moved from merge-space to regular collection storage using the 2-phase commit feature of GbMoveFile.cpp, and finally reads are done from the finished file.

Details:
  RdbBase: Use MergeSpaceCoordinator and merge space for temporary target merge file.
  RdbBase: better cleanup of crashed merges
  RdbBase: more mutex locing while manipulatin m_fileInfo array
  RdbBase: keep track of thraeds/jobs
  RdbMerge: ditto
  RdbMerge: Dont call file->chopHead()
  Msg5/Msg3: no more "compensate for merge" flag
  Msg3: Skip over RdbBase files that have reads disallowed