Removed commented-out signal handlers for SIGSEGV/SIGILL/SIGFPE/SIGBUS
Original code caught those signals and tried to save (potentially corrupted) data to disk before trying to get a stack backtrace and log that.
it was commented out in an earlier commit because saving potentially corrupt data is not exactly a good idea, and the get-stack-backtre was less reliable than gdb, so the end result was often potentially corrutp data and a useless or non-existing stack backtrace.