Check for erros when finishing merge
If any of the operations (unlinking source files, moving/copying resulting file) then shut down, and wait for human investigation/intervention.
This is needed because if eg. open(RD_ONLY) on an existing file fails with eg. EINVAL then we hav eno recovery strategy because we it is unclear why it failed.