Remoed clock synchronization, time adjustment etc.
The assumption is that on modern systems NTP or similar is used on the hosts so GB doesn't have to do that job. If clocks actualyl do drift apart then no major harm is donw, although we may want to detect that and shut down if clock differences are too great.