Dropped PingInfo + PingServer entierely.
The remaining functionality of the pingserver was to ping normalt and "shotgun" IPs to determine the fastests and working network path. This is better done at L2/L3 or OS level - not in the application.

The email-sending functionality was dropped too. We recommend using 3rd-party tools for monitoring and statistics-collection/triggering.