Moved allocation of working memory from msg39 to intersect
The m_top_scoringdocidibuf member was only used during intersection. We handle memory allocation errors in threads fine now, so the allocation was moved from the msg39 coordinator thread to the intersection thread. This should keep the memory nearer to where it will be used, and only allocated while it is being used.