Fix allocation of working-area in PosdbTable::intersectLists10_r()
The local arrays were allocated for num-query-terms but what was really needed was num-query-term-infos. Changed to plain std::vector<> arrays and call to intersectLists() wrapped for cathcing std::bad_alloc. 'scoreMatrix' is still kept as a lineraized verison of a 2-dimensional matrix.

No changed all the way through so there a still some "&(vector[0])" expressions.