Removed "hack of confusion" in PosdbTable.cpp
The "hack" was that if the query used two words eg "rabbit" and "foot" and we're looking at the bigram for that, and the posdb entry/document used two fwords too then don't add the bigram entries to the mergebuf.
Unclear why that logic was there because neither ranking nor performance seem to be affected by it.