Don't set S(ynonym) bit in bigram posdb entries.
The S bit was set for bigrams so the scoring later on could perform som non-obvious logic on it (see commit 9f072606d3b12f14d723bf4f34e9d93a7171394b). But that logic is gone and there is no need to overload that bit in the entriy two two meanings.