Use correct term-freq-weight for bigrams and synonyms
When search for "abc xyz" the term-freq-weight used for scoring should be the term-freq-weight of the term from which the posdb-list came, and not the base term. Previously the term-freq-weight used was the same as the base term, so eg. synonyms and bigrams term-freq-weight dependon which base term was being processed.