posdbtable: allow multiple sublists per wordpos
mergeTermSubListsForDocId() used to prevent multiple entries with the same wordpos. That resitrction has been removed and there seem to be no ill effect. It allows Query and setQueryTermInfo() to add term (sub) lists in any order, and also solves a bug when sarching for two terms and the second term is a high-freq-term og querystopword.