bugfix: if last match in a document was the best then sometimes the document was eliminated
findMinTermPairScoreInWindow() would find the best match, while getMinTermPairScoreSlidingWindow() woudl set xpos to NULL when that query-term-info list was exhausted. but if the remaining terms had boosts big enough and were the best match the m_bestMinTermPairWindowPtrs[] would be set to xpos, but now with a NULL ptr in it. getTermPairScoreForAny() treated a NULL presence as score = 0.0, and thus the document was eliminated.